requirements.txt
annotated-types==0.7.0
anthropic==0.43.1
anyio==4.8.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
google-auth==2.37.0
griffe==1.5.5
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpath-python==1.0.6
logfire-api==3.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
mistralai==1.4.0
mypy-extensions==1.0.0
openai==1.59.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-ai==0.0.19
pydantic-ai-slim==0.0.19
pydantic-graph==0.0.19
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
